Add tests for project validation
