+++ tests/unit/data/invalid/proj-missing-flow/pkg_a/metadata/Flow/test_missing_flow.json
@@ -1,0 +1,56 @@
+{
+    "Flow": {
+        "test_missing_flow": {
+            "Container": {
+                "device_api": {
+                    "ContainerFlow": "device_api",
+                    "ContainerLabel": "device_api",
+                    "Visualization": {
+                        "x": {
+                            "Value": 196
+                        },
+                        "y": {
+                            "Value": 358
+                        }
+                    }
+                },
+                "tugbot": {
+                    "ContainerFlow": "tugbot",
+                    "ContainerLabel": "tugbot",
+                    "Visualization": {
+                        "x": {
+                            "Value": 195
+                        },
+                        "y": {
+                            "Value": 87
+                        }
+                    }
+                }
+            },
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_missing_flow",
+            "LastUpdate": {
+                "date": "07/07/2023 at 09:28:39",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "c4087d62-e7f1-4d45-b1a8-caeb5d78137c": {
+                    "Dependency": 0,
+                    "From": "start/start/start",
+                    "To": "device_api__get_task/start/in"
+                },
+                "fd98516b-a4c3-4510-89a3-3d3959542148": {
+                    "Dependency": 0,
+                    "From": "start/start/start",
+                    "To": "tugbot__urdf/start/in"
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
