+++ tests/unit/data/invalid/proj-missing-node-instance/pkg_a/manifest.txt
@@ -1,0 +1,1 @@
+Flow:test_missing_flow
