+++ tests/unit/data/invalid/proj-missing-node-instance/pkg_a/metadata/Flow/test_missing_flow.json
@@ -1,0 +1,25 @@
+{
+    "Flow": {
+        "test_missing_flow": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_missing_flow",
+            "LastUpdate": {
+                "date": "07/07/2023 at 09:28:39",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "c4087d62-e7f1-4d45-b1a8-caeb5d78137c": {
+                    "Dependency": 0,
+                    "From": "start/start/start",
+                    "To": "non_existing_instance/start/in"
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
