+++ tests/unit/data/invalid/proj-missing-node/pkg_a/metadata/Flow/test_missing_node.json
@@ -1,0 +1,54 @@
+{
+    "Flow": {
+        "test_missing_node": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_missing_node",
+            "LastUpdate": {
+                "date": "07/11/2022 at 11:34:56",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "8395d1a2-af21-4b12-a1a3-e2dfbf7c198f": {
+                    "From": "start/start/start",
+                    "To": "test_any/sub/in"
+                },
+                "20893b58-911b-470d-9306-1e4ac32b76d1": {
+                    "From": "start/start/start",
+                    "To": "test_something/sub/in"
+                }
+            },
+            "NodeInst": {
+                "test_any": {
+                    "NodeLabel": "test_any",
+                    "Template": "test_any",
+                    "Visualization": {
+                        "x": {
+                            "Value": 0.010400000000000001
+                        },
+                        "y": {
+                            "Value": 0.005866666666666667
+                        }
+                    }
+                },
+                "test_something": {
+                    "NodeLabel": "test_any",
+                    "Template": "test_any",
+                    "Visualization": {
+                        "x": {
+                            "Value": 0.010400000000000001
+                        },
+                        "y": {
+                            "Value": 0.005866666666666667
+                        }
+                    }
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
