+++ tests/unit/data/invalid/proj-missing-port/pkg_a/manifest.txt
@@ -1,0 +1,2 @@
+Flow:test_transition_to_any
+Node:dependency
