+++ tests/unit/data/invalid/proj-missing-port/pkg_a/metadata/Flow/test_missing_port.json
@@ -1,0 +1,39 @@
+{
+    "Flow": {
+        "test_missing_port": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_missing_port",
+            "LastUpdate": {
+                "date": "07/07/2023 at 15:03:39",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "a42d10cf-32c4-40a4-a95d-812649d56810": {
+                    "Dependency": 0,
+                    "From": "start/start/start",
+                    "To": "dependency/entry/in"
+                }
+            },
+            "NodeInst": {
+                "dependency": {
+                    "NodeLabel": "dependency",
+                    "Template": "dependency",
+                    "Visualization": {
+                        "x": {
+                            "Value": 178
+                        },
+                        "y": {
+                            "Value": 96
+                        }
+                    }
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
