+++ tests/unit/data/invalid/proj-missing-port/pkg_a/metadata/Node/dependency.json
@@ -1,0 +1,40 @@
+{
+    "Node": {
+        "dependency": {
+            "Info": "",
+            "Label": "dependency",
+            "LastUpdate": "10/06/2020 at 16:41:57",
+            "Launch": true,
+            "PackageDepends": "",
+            "Path": "",
+            "Persistent": false,
+            "PortsInst": {
+                "dependency": {
+                    "In": {
+                        "in": {
+                            "Message": "movai_msgs/Dependency"
+                        }
+                    },
+                    "Message": "Dependency",
+                    "Package": "movai_msgs",
+                    "Template": "MovAI/Dependency"
+                },
+                "depends": {
+                    "Message": "Dependency",
+                    "Out": {
+                        "out": {
+                            "Message": "movai_msgs/Dependency"
+                        }
+                    },
+                    "Package": "movai_msgs",
+                    "Template": "MovAI/Depends"
+                }
+            },
+            "Remappable": true,
+            "Type": "MovAI/State",
+            "User": "movai",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
