+++ tests/unit/data/invalid/proj-non-matching-ports/pkg_a/manifest.txt
@@ -1,0 +1,10 @@
+Flow:test_transition_to_dependency
+Node:dependency
+
+Flow:test_transition_to_ros
+Node:test_ros
+
+Flow:test_pub_ros_to_sub_ros
+
+Flow:test_transition_to_any
+Node:test_any
