+++ tests/unit/data/invalid/proj-non-matching-ports/pkg_a/metadata/Flow/test_pub_ros_to_sub_ros.json
@@ -1,0 +1,38 @@
+{
+    "Flow": {
+        "test_pub_ros_to_sub_ros": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_pub_ros_to_sub_ros",
+            "LastUpdate": {
+                "date": "04/11/2022 at 18:41:10",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "ca35667e-8e58-4c71-8973-245da65dbe0b": {
+                    "From": "ros/pub_empty/out",
+                    "To": "ros/sub/in"
+                }
+            },
+            "NodeInst": {
+                "ros": {
+                    "NodeLabel": "ros",
+                    "Template": "test_ros",
+                    "Visualization": {
+                        "x": {
+                            "Value": 0.0126
+                        },
+                        "y": {
+                            "Value": 0.004933333333333334
+                        }
+                    }
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
