+++ tests/unit/data/invalid/proj-non-matching-ports/pkg_a/metadata/Flow/test_transition_to_any.json
@@ -1,0 +1,38 @@
+{
+    "Flow": {
+        "test_transition_to_any": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_transition_to_any",
+            "LastUpdate": {
+                "date": "07/11/2022 at 11:34:56",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "8395d1a2-af21-4b12-a1a3-e2dfbf7c198f": {
+                    "From": "start/start/start",
+                    "To": "test_any/sub/in"
+                }
+            },
+            "NodeInst": {
+                "test_any": {
+                    "NodeLabel": "test_any",
+                    "Template": "test_any",
+                    "Visualization": {
+                        "x": {
+                            "Value": 0.010400000000000001
+                        },
+                        "y": {
+                            "Value": 0.005866666666666667
+                        }
+                    }
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
