+++ tests/unit/data/invalid/proj-non-matching-ports/pkg_a/metadata/Flow/test_transition_to_ros.json
@@ -1,0 +1,38 @@
+{
+    "Flow": {
+        "test_transition_to_ros": {
+            "Description": "",
+            "ExposedPorts": {},
+            "Info": "",
+            "Label": "test_transition_to_ros",
+            "LastUpdate": {
+                "date": "04/11/2022 at 18:05:50",
+                "user": "movai@internal"
+            },
+            "Layers": {},
+            "Links": {
+                "20893b58-911b-470d-9306-1e4ac32b76d1": {
+                    "From": "start/start/start",
+                    "To": "ros/sub/in"
+                }
+            },
+            "NodeInst": {
+                "ros": {
+                    "NodeLabel": "ros",
+                    "Template": "test_ros",
+                    "Visualization": {
+                        "x": {
+                            "Value": 0.0152
+                        },
+                        "y": {
+                            "Value": 0.0128
+                        }
+                    }
+                }
+            },
+            "User": "",
+            "Version": "",
+            "VersionDelta": {}
+        }
+    }
+}
